ui: refactor date constant usage
formatTimeAgo had a really weird pattern where the caller had to pass
in the values of constants, that always needed to be the same. I can't
see any good reason for this.

Move the constants to a more appropriate package and adjust imports
accordingly.
