+++ ui/ops/src/components/download/download.js
@@ -1,6 +1,7 @@
 import {timestampToDate} from '@/api/convpb.js';
 import {getDownloadDevicesUrl} from '@/api/ui/devices.js';
-import {MINUTE} from '@/components/now.js';
+
+import {MINUTE} from '@/util/date.js';
 import {computed, onScopeDispose, ref, toValue, watch} from 'vue';
 
 /**
