+++ ui/ops/src/components/now.js
@@ -1,10 +1,5 @@
+import {MINUTE} from '@/util/date.js';
 import {onScopeDispose, ref, toValue} from 'vue';
-
-export const MILLISECOND = 1;
-export const SECOND = 1000 * MILLISECOND;
-export const MINUTE = 60 * SECOND;
-export const HOUR = 60 * MINUTE;
-export const DAY = 24 * HOUR;
 
 /**
  *
