+++ ui/ops/src/dynamic/widgets/power-history/PowerHistoryGraph.vue
@@ -28,11 +28,12 @@ import LineChart from '@/components/charts/LineChart.vue';
 import ChartTooltip from '@/components/charts/ChartTooltip.vue';
 import {useExternalTooltip} from '@/components/charts/plugins.js';
 import {triggerDownload} from '@/components/download/download.js';
-import {HOUR, MINUTE, useNow} from '@/components/now.js';
+import {useNow} from '@/components/now.js';
 import useTimePeriod from '@/composables/useTimePeriod.js';
 import PowerHistoryGraphOptionsMenu from '@/dynamic/widgets/power-history/PowerHistoryGraphOptionsMenu.vue';
 import useMeterHistory from '@/dynamic/widgets/power-history/useMeterHistory.js';
 import {useCarbonIntensity} from '@/stores/carbonIntensity.js';
+import {HOUR, MINUTE} from '@/util/date.js';
 import {format} from '@/util/number.js';
 import {computed, ref, watch} from 'vue';
 import {useTheme} from 'vuetify';
